Range Locking: make GetLockStatusData report locks in STO mode
(Implement this in locktree:dump_locks)

Before this patch, GetLockStatusData would return empty list when
the lock tree is in STO (Single Transaction Optimization mode.
This is confusing.

Note: some MTR testcases have workarounds for this: they get a lock in
another connection to make sure we're not in STO-mode.
They will still have to do that, because STO-mode may report different
set of locks (e.g. the locks can overlap with one another).