Remove PessimisticTransactionDB::range_lock_mgr_
It is not needed since PessimisticTransactionDB code doesn't call
range_lock_mgr_->UnLockAll() anymore.

If we don't need range_lock_mgr_, we also don't need
PessimisticTransactionDB::init_lock_manager() to set it.

We can use NewLockManager() but need it to return shared_ptr since
sometimes it creates an object it returns and sometimes it just
takes it from TransactionDBOptions.